# swagger/openapi.yaml
...
openapi: 3.0.0
servers:
- url: http://localhost:3000/
info:
  description: 'This is a sample E-Commerce app.  This project requires that you build a fully-functioning e-commerce application REST API that allows users to perform various CRUD operations such as registering an account, browsing products for sale, etc.'
  version: 1.0.0
  title: E-Commerce REST API application
  termsOfService: https://www.codecademy.com/about
  contact:
    email: easoftware@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/orders":
    post:
      tags:
      - orders
      summary: Create order
      description: 'Done by the user'
      operationId: createOrder
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Orders"
        '400':
          description: required value missing, invalid data
        '409':
          description: order_id or product_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/OrdersPost"
    get:
      tags:
      - orders
      summary: Gets all orders
      description: ''
      operationId: getAllOrders
      responses:
        '200':
          description: successful opertion (No orders rows is OK, array is empty)
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Orders"
  "/orders/{id}":  
    get:
      tags:
      - orders
      summary: Get order by order id
      description: ''
      operationId: getOrderById
      parameters:
      - name: id
        in: path
        description: 'The order id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Orders"
        '404':
          description: Invalid Order ID supplied, Order not found
    put:
      tags:
      - orders
      summary: update order by order id
      description: 'This can only be done by the logged in user.'
      operationId: updateOrderById
      parameters:
      - name: id
        in: path
        description: 'The order id of the order that needs to be updated. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Orders"
        '400':
          description: required value missing
        '404':
          description: Invalid Order ID supplied, Order not found
        '409':
          description: user_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/OrdersPut"
    delete:
      tags:
      - orders
      summary: Delete order
      description: This can only be done by the app or Admin.
      operationId: deleteOrder
      parameters:
      - name: id
        in: path
        description: 'The order id that needs to be deleted. In testing, use before() to generate cart to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted order - order's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '400': 
          description: unknown error
        '404':
          description: Order not found
        '409':
          description: Cannot delete - constraint error          
  "/orders/{id}/items":
    post:
      tags:
      - orders
      summary: Create order item
      description: 'Done when user checks out.'
      operationId: createOrderItem
      parameters:
      - name: id
        in: path
        description: 'The order id link the order item. Use 1 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order_items"
        '400':
          description: product id not valid, required value missing
        '404':
          description: order not found
        '409':
          description: order_id or product_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPost"
    get:
      tags:
      - orders
      summary: Get order items by order id
      description: ''
      operationId: getOrderItemsById
      parameters:
      - name: id
        in: path
        description: 'The order id for items to be fetched. Use 1 for testing.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Order_items"
        '404':
          description: Invalid Cart ID supplied, Cart not found
  "/orders/items/{itemId}":
    get:
      tags:
      - orders
      summary: Get order item by order item id
      description: ''
      operationId: getOrderItemById
      parameters:
      - name: itemId
        in: path
        description: 'The order item id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order_items"
        '404':
          description: Invalid Order Item ID supplied, order item not found
    put:
      tags:
      - orders
      summary: Get order item by order item id
      description: ''
      operationId: getOrderItemByOrderId
      parameters:
      - name: itemId
        in: path
        description: 'The order item id for items to be fetched. Use 2 for testing.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order_items"
        '400':
          description: product id not valid, required value missing
        '404':
          description: Invalid Order Item ID supplied, Order Item not found
        '409':
          description: product_id violates foreign key constraint
      requestBody:
        "$ref": "#/components/requestBodies/CartItemsPut"
    delete:
      tags:
      - orders
      summary: Delete order item
      description: Done by the user
      operationId: deleteOrderItem
      parameters:
      - name: itemId
        in: path
        description: 'The order item id that needs to be deleted. In testing, use before() to generate order item to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted order item - order item's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Order item not found, invalid order item id
components:
  schemas:
...
    Orders:
      type: object
      properties:
        id:
          type: integer
          format: int32
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        status:
          type: string
        total_price:
          type: number
          format: Decimal
        user_id:
          type: integer
          format: int32
    Order_items:
      type: object
      properties:
        id:
          type: integer
          format: int32
        order_id:
          type: integer
          format: int32
        product_id:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        price_unit:
          type: number
          format: Decimal
...
  requestBodies:
    CartItemsPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              cart_id:
                type: integer
                format: int32
              product_id:
                type: integer
                format: int32
              quantity:
                type: integer
                format: int32
    CartItemsPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              product_id:
                type: integer
                format: int32
              quantity:
                type: integer
                format: int32
...
    OrdersPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              created:
                type: string
                format: date-time
              total_price:
                type: number
                format: Decimal
              used_id:
                type: integer
                format: int32
    OrdersPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              modified:
                type: string
                format: date-time
              status:
                type: string
              total_price:
                type: number
                format: Decimal
              used_id:
                type: integer
                format: int32
...